add component container and card
