+++ src/App.tsx
@@ -18,7 +18,6 @@ import DownloadApp from "./pages/buyers/DownloadApp";
 import Profile from "./pages/buyers/Profile";
 import Data from "./pages/profile/Data";
 import Reset from "./pages/profile/Reset";
-import DumyHome from "./pages/DumyHome";
 
 const router = createBrowserRouter([
   {
